Broke Roll class out into a file
